# optmize baichuan in pytorch engine